#23 Update create_request Endpoint
This commit modifies the endpoint to write new requests to the database. This allows us to retrieve the automatically generated request_id to include it in the Job paramaters, improving consistency when handling the prediction Jobs in a separate thread. This also allows us to return the request_id to the user, enabling the client to efficiently poll for updates for a given request.