#23 Add More Unit Tests For Prediction Preprocessing
This commit adds more unit tests that assert the correctness of the helper functions that are used by the prediction manager to preprocess the jobs (which contain the user data) before running inference. A test asserting the assumption that the helper functions preserve the ordering of data has also been added. Furtermore, this commit deletes some tests that were asserting behaviour in the helper functions where they are able to operate on illegal values without throwing errors. This is to prevent confusion if the application were to crash from such values being passed to the prediction manager in which case these tests would pass, misleadingly suggesting that the illegal values are expected.